+++ libs/cpp/parse/parser.h
@@ -113,7 +113,7 @@ public:
     const std::string& group() const { return _event_definition.group_name; }
     const std::string& type() const { return _event_definition.type; }
 
-    int numArguments() const { return _event_definition.arguments.size(); }
+    int numArguments() const { return static_cast<int>(_event_definition.arguments.size()); }
     const EventArgumentDefinition& argument(int index) const { return _event_definition.arguments[index]; }
     Argument argumentValue(int index) const;
     uint64_t argumentValueInt(int index) const;
